+++ app/Http/Controllers/API/apiController.php
@@ -3,13 +3,11 @@
 namespace App\Http\Controllers\API;
 
 use App\Http\Controllers\Controller;
-use App\Models\Data;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Hash;
 use App\Models\GreenHouse;
 use App\Models\Sensor;
 use App\Models\Zone;
-use App\Models\Notification;
 use Illuminate\Http\Request;
 use Psy\Util\Json;
 use Illuminate\Support\Facades\Auth;
@@ -260,134 +258,5 @@ class apiController extends Controller
             return $datas[0]->data;
         }
     }
-    //Posting data in database
-    public function postData(Request $request){
-        $user = Auth::user();
-        try {
-            //Getting the ID of the company
-            $company = DB::table('tblGreenHouse')
-                ->leftjoin('tblZone','tblGreenHouse.idGreenHouse','=','tblZone.idGreenHouse')
-                ->leftjoin('tblSensor','tblZone.idZone','=','tblSensor.idZone')
-                ->select('tblGreenHouse.idCompany')
-                ->where('tblSensor.idSensor','=',$request['sensor'])
-                ->pluck('idCompany');
-
-
-            if($company[0] == $user['idCompany']){
-                //The captor is owned by the company, so it's good
-                $data = new Data;
-
-                $data->data = $request['data'];
-                $data->idSensor = $request['sensor'];
-                verifyData($data);
-                $data->save();
-
-                $response = 'Accepted';
-                return response($response, 201);
-            }
-            else{
-                //Not owned by the company
-                $response = 'This captor is not owned by the company';
-                return response($response, 401);
-            }
-        }
-        catch(\Illuminate\Database\QueryException $ex){
-            $response = 'An error occurred';
-            return response($response, 400);
-        }
-    }
-
-    // Verify data before post and create a notification if data is too high/too low
-    public function VerifyData($data){
-
-        try {
-
-            // Looking for the sensor if it is the one of temperature
-            $typeData = DB::table('tblSensor')
-            ->select('tblSensor.typeData')
-            ->where('tblSensor.idSensor','=',$data['sensor'])
-            ->pluck('typeData');
-
-            // Verify if data sent is in a correct temperature
-            if($typeData[0] == "humidite"){
-
-                $notification = Notification::find($id);
-                $status = $notification->alerteStatus;
-
-                //Status 0 = On Fire
-                //Status 1 = idle
-
-                // Look if there is a fire
-                if($data > 50){
-                    // Update the database if something has changed
-                    if($status == 1){
-                        $notification->description = '7.8/10 too much water -IGN';
-                        $notification->alerteStatus = 0;
-                        $notification->save();
-                    }         
-                }
-                // No fire
-                else{
-                    //Previously on fire
-                    if($status == 0){
-                        $notification->description = 'Everything is fine :)';
-                        $notification->alerteStatus = 1;
-                        $notification->save();
-                    }
-                }        
-            } 
-        }
-        catch(\Illuminate\Database\QueryException $ex){
-            // Return the exception and the error 
-            $response = 'An error occurred';
-            return response($response, 400);
-        }
-    }
-
-
-    //Returning if you need to water the plant or not
-    public function getWater(Request $request, $idZone){
-
-        //todo - Api call to check how much water the zone need
-
-        $user = Auth::user();
-        try{
-            $zone = Zone::find($idZone);
-            //Getting the ID of the company
-            $company = DB::table('tblGreenHouse')
-                ->leftjoin('tblZone','tblGreenHouse.idGreenHouse','=','tblZone.idGreenHouse')
-                ->select('tblGreenHouse.idCompany')
-                ->where('tblZone.idZone','=',$idZone)
-                ->pluck('idCompany');
-
-            if($company[0] == $user['idCompany']){
-
-                //The zone is owned by the company, so it's good
-                $response = [
-                    'water' => $zone->water,
-                    'quantity' => 300
-                ];
-
-                //update the water to false
-                if($zone->water == 0){
-                    $zone->water = 1;
-                    $zone->save();
-                }
-
-
-                return response($response, 201);
-            }
-            else{
-                //Not owned by the company
-                $response = 'This captor is not owned by the company';
-                return response($response, 401);
-            }
-        }
-        catch(\Illuminate\Database\QueryException $ex){
-            $response = 'An error occurred';
-            return response($response, 400);
-        }
-    }
-    //Returning if you need to water the plant or not
 
 }
